refactor: change GithubPullsItem to inner class
