--- src/main/java/com/github/hcsp/http/GithubPullsItem.java
@@ -1,45 +1,0 @@
-package com.github.hcsp.http;
-
-public class GithubPullsItem {
-
-    public static class User {
-        private String login;
-
-        public String getLogin() {
-            return login;
-        }
-
-        public void setLogin(String login) {
-            this.login = login;
-        }
-    }
-
-    private int number;
-    private String title;
-
-    private User user;
-
-    public int getNumber() {
-        return number;
-    }
-
-    public void setNumber(int number) {
-        this.number = number;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-}
